Add new interfaces and update existing ones

+++ projects/market/src/lib/types/GDataTypes.sol
@@ -3,14 +3,14 @@ pragma solidity ^0.8.20;
 
 library GDataTypes {
     struct ValidParams {
-        uint128 sizeDelta;
-        uint128 globalLongSizes;
-        uint128 globalShortSizes;
-        uint128 userLongSizes;
-        uint128 userShortSizes;
-        uint128 marketLongSizes;
-        uint128 marketShortSizes;
-        uint128 aum;
+        uint256 sizeDelta;
+        uint256 globalLongSizes;
+        uint256 globalShortSizes;
+        uint256 userLongSizes;
+        uint256 userShortSizes;
+        uint256 marketLongSizes;
+        uint256 marketShortSizes;
+        uint256 aum;
         bool isLong;
         uint16 market;
     }
